# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News portal",
        "description": "Documentation for the NewsAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/favourites": {
            "get": {
                "tags": [
                    "Favourites"
                ],
                "summary": "Get a listing of user's favourite articles.",
                "description": "Returns a list of articles that the user has marked as favourites.",
                "operationId": "favourites.index",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Article Title"
                                            },
                                            "url": {
                                                "type": "string",
                                                "example": "https://example.com/article"
                                            },
                                            "author": {
                                                "type": "string",
                                                "example": "Author Name"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Article Description"
                                            },
                                            "image_url": {
                                                "type": "string",
                                                "example": "https://example.com/image.jpg"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2023-12-01T12:34:56Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found - The requested resource could not be found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Favourites"
                ],
                "summary": "Store a newly created favourite article in databes.",
                "description": "Returns redirect to article.index.",
                "operationId": "favourites.store",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not Found - The requested resource could not be found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/favourites{favourite_id}": {
            "delete": {
                "tags": [
                    "Favourites"
                ],
                "summary": "Remove the specified favourite article from database.",
                "description": "Returns redirect to article.index.",
                "operationId": "favourites.destroy",
                "parameters": [
                    {
                        "name": "favourite_id",
                        "in": "path",
                        "description": "ID of the favorite article to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not Found - The requested resource could not be found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}